requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
httplib2==0.19.1
idna==2.10
mysql-connector-python==8.0.24
numpy==1.20.2
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
urllib3==1.26.4
